Update ElectricityPrice tests to use BigDecimal

+++ src/test/java/com/martinbg/electricityapi/model/ElectricityPriceTest.java
@@ -1,14 +1,17 @@
 package com.martinbg.electricityapi.model;
 
 import org.junit.jupiter.api.Test;
+
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 class ElectricityPriceTest {
 
     @Test
     void testElectricityPriceConstructor() {
-        double daily = 0.17514;
-        double night = 0.07546;
+        BigDecimal daily = BigDecimal.valueOf(0.17514);
+        BigDecimal night = BigDecimal.valueOf(0.07546);
 
         ElectricityPrice electricityPrice = new ElectricityPrice(daily, night);
 
@@ -18,9 +21,9 @@ class ElectricityPriceTest {
 
     @Test
     void testSetters() {
-        ElectricityPrice electricityPrice = new ElectricityPrice(0.1, 0.2);
-        double newDaily = 0.25;
-        double newNight = 0.15;
+        ElectricityPrice electricityPrice = new ElectricityPrice(BigDecimal.valueOf(0.1), BigDecimal.valueOf(0.2));
+        BigDecimal newDaily = BigDecimal.valueOf(0.25);
+        BigDecimal newNight = BigDecimal.valueOf(0.15);
 
         electricityPrice.setDaily(newDaily);
         electricityPrice.setNight(newNight);
